+++ src/Tests/DependencyInjection/SurfnetYubikeyApiClientExtensionTest.php
@@ -1,8 +1,11 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace Surfnet\YubikeyApiClientBundle\Tests\DependencyInjection;
 
-use PHPUnit_Framework_TestCase as TestCase;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
+use PHPUnit\Framework\TestCase;
 use Surfnet\YubikeyApiClientBundle\Tests\TestKernel;
 use Symfony\Component\DependencyInjection\Container;
 
@@ -12,13 +15,14 @@ class SurfnetYubikeyApiClientExtensionTest extends TestCase
      * @test
      * @group DependencyInjection
      */
-    public function verification_service_can_be_loaded()
+    public function verificationServiceCanBeLoaded()
     {
         $kernel = new TestKernel('test', true);
         $kernel->boot();
 
         /** @var Container $container */
         $container = $kernel->getContainer();
+        $this->expectNotToPerformAssertions();
         $container->get('surfnet_yubikey_api_client.verification_service');
     }
 }