+++ CHANGELOG.md
@@ -1,9 +1,24 @@
 # Changelog
 
+## 4.2.0  
+Drop support for SF 4
+Changed to GHA
+Added extra QA tools
+
+## 4.1.1
+
+Do not pin dependency versions
+
+## 4.1.0
+
+Upgrade yubikey-api-client to use new YK api endpoints
+
 ## 3.0.1
+
 Remove RMT and Ibuildings QA tools from the project.
 
 ## Older versions
+
 Older entries can be found on the projects github page.
 See https://github.com/SURFnet/yubikey-api-client-bundle/releases